[level 1] Title: 두 정수 사이의 합, Time: 0.01 ms, Memory: 4.89 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.cpp
@@ -6,10 +6,10 @@ using namespace std;
 long long solution(int a, int b) {
     long long answer = 0;
     
-    long n = a+b;
-    long cnt = a > b ? a - b + 1 : b - a + 1;
+    int n = a+b;
+    int cnt = a > b ? a - b + 1 : b - a + 1;
     
-    answer = cnt*n/2;
+    answer = (long)cnt*n/2;
     
     return answer;
 }
